requirements.txt
# PostgreSQL
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.4
numpy==1.26.0

# Technical indicators
ta==0.11.0

# Machine Learning
tensorflow==2.16.1
scikit-learn==1.3.2
xgboost==2.0.3
optuna==3.5.0
imbalanced-learn==0.12.0

# Visualization
matplotlib==3.8.2

# HTTP requests for Bybit API
requests==2.31.0
pybit==5.6.2

# Progress bars
tqdm==4.66.1

# Date/time handling
python-dateutil==2.8.2

# Configuration
pyyaml==6.0.1